Replace dictionary with elif ladder
When using a dictionary to return the value based on generic_type,
all the cases are executed when creating the dictionary (i.e. all
the objects will be created and the right one will be picked in
the end). Which is obviously a waste (and gets buggy because of
unintentional calls into _new_sensor). Replace the dictionary with
a elif ladder. This is probably a bit more characters, but worth
the performance improvement.